finished the Dock page with all ships, updated some starting resource needs of ships, and others. regenerated the statistics.

+++ omega-model/src/main/java/pl/omega/model/objects/Defending.java
@@ -2,8 +2,9 @@ package pl.omega.model.objects;
 
 public interface Defending {
 
+	// TODO Adam Puchalski - Apr 29, 2012 - maybe a datatype change is needed?
 	public int getStructurePoints ();
-	public int getShield();
-	public int getAttack();
+	public double getShield();
+	public double getAttack();
 	
 }
